Move structs Message, Payload... from process(process) to process(io)
